ViewportDesc and ScissorsDesc introduced

+++ DopeEngine/src/Engine/Graphics/RenderPass/ScissorsDesc.h
@@ -5,13 +5,13 @@ namespace DopeEngine
 {
 	struct DOPE_ENGINE_API ScissorsDesc
 	{
-		ScissorsDesc(const float offsetX,const float offsetY,const float width,const float height) : OffsetX(offsetX),OffsetY(offsetY),Width(width),Height(height) {}
+		ScissorsDesc(const unsigned int offsetX,const unsigned int offsetY,const unsigned int width,const unsigned int height) : OffsetX(offsetX),OffsetY(offsetY),Width(width),Height(height) {}
 		ScissorsDesc() : OffsetX(0),OffsetY(0),Width(0),Height(0) {}
 		~ScissorsDesc() = default;
 
-		float OffsetX;
-		float OffsetY;
-		float Width;
-		float Height;
+		unsigned int OffsetX;
+		unsigned int OffsetY;
+		unsigned int Width;
+		unsigned int Height;
 	};
 }
